+++ xlisp/xlprin.c
@@ -245,7 +245,7 @@ LOCAL void putqstring(LVAL fptr, LVAL str)
 void putatm(LVAL fptr, char *tag, LVAL val)
 {
     sprintf(buf,"#<%s: #",tag); xlputstr(fptr,buf);
-    sprintf(buf,AFMT,(int)val); xlputstr(fptr,buf);
+    sprintf(buf,AFMT,(long unsigned int)val); xlputstr(fptr,buf);
     xlputc(fptr,'>');
 }
 
@@ -254,7 +254,7 @@ LOCAL void putsubr(LVAL fptr, char *tag, LVAL val)
 {
     sprintf(buf,"#<%s-%s: #",tag,funtab[getoffset(val)].fd_name);
     xlputstr(fptr,buf);
-    sprintf(buf,AFMT,(int)val); xlputstr(fptr,buf);
+    sprintf(buf,AFMT,(long unsigned int)val); xlputstr(fptr,buf);
     xlputc(fptr,'>');
 }
 
@@ -267,7 +267,7 @@ LOCAL void putclosure(LVAL fptr, LVAL val)
     else
         strcpy(buf,"#<Closure: #");
     xlputstr(fptr,buf);
-    sprintf(buf,AFMT,(int)val); xlputstr(fptr,buf);
+    sprintf(buf,AFMT,(long unsigned int)val); xlputstr(fptr,buf);
     xlputc(fptr,'>');
 /*
     xlputstr(fptr,"\nName:   "); xlprint(fptr,getname(val),TRUE);
